NUMBERS-188 made javadoc changes
