+++ commons-numbers-complex/src/main/java/org/apache/commons/numbers/complex/ComplexFunctions.java
@@ -25,10 +25,6 @@ import java.util.function.DoubleUnaryOperator;
  * complex number \( a + ib \), \( a \) is called the <em>real part</em> and
  * \( b \) is called the <em>imaginary part</em>.
  *
- * <p>This class is immutable, it is to change from an OO representation of a Complex number
- * to a representation as its real and imaginary parts. All arithmetic will create a new instance for the
- * result.</p>
- *
  * <p>Arithmetic in this class conforms to the C99 standard for complex numbers
  * defined in ISO/IEC 9899, Annex G. Methods have been named using the equivalent
  * method in ISO C99. The behavior for special cases is listed as defined in C99.</p>
@@ -73,7 +69,8 @@ public final class ComplexFunctions {
     private static final double LOG10_2 = Math.log10(2);
     /** {@code sqrt(2)}. */
     private static final double ROOT2 = 1.4142135623730951;
-
+    /** The bit representation of {@code -0.0}. */
+    private static final long NEGATIVE_ZERO_LONG_BITS = Double.doubleToLongBits(-0.0);
     /** 54 shifted 20-bits to align with the exponent of the upper 32-bits of a double. */
     private static final int EXP_54 = 0x36_00000;
     /** Represents an exponent of 500 in unbiased form shifted 20-bits to align with the upper 32-bits of a double. */
@@ -87,6 +84,11 @@ public final class ComplexFunctions {
     private static final double TWO_POW_600 = 0x1.0p+600;
     /** 2^-600. */
     private static final double TWO_POW_NEG_600 = 0x1.0p-600;
+
+    /** The multiplier used to split the double value into hi and low parts. This must be odd
+     * and a value of 2^s + 1 in the range {@code p/2 <= s <= p-1} where p is the number of
+     * bits of precision of the floating point number. Here {@code s = 27}.*/
+    private static final double MULTIPLIER = 1.34217729E8;
 
     /**
      * Private constructor for utility class.
@@ -111,8 +113,8 @@ public final class ComplexFunctions {
      * <p>This method is called by all methods that require the absolute value of the complex
      * number, e.g. abs(), sqrt() and log().
      *
-     * @param real Real part \( a \) of the complex number \(a +ib \).
-     * @param imaginary Imaginary part \( b \) of the complex number \(a +ib \).
+     * @param real Real part \( a \) of the complex number \( (a +ib) \).
+     * @param imaginary Imaginary part \( b \) of the complex number \( (a +ib) \).
      * @return The absolute value.
      */
     public static double abs(double real, double imaginary) {
@@ -141,8 +143,8 @@ public final class ComplexFunctions {
      * in calculating the returned value using the {@code atan2(y, x)} method for complex
      * \( x + iy \).
      *
-     * @param r Real part \( a \) of the complex number \(a +ib \).
-     * @param i Imaginary part \( b \) of the complex number \(a +ib \).
+     * @param r Real part \( a \) of the complex number \( (a +ib) \).
+     * @param i Imaginary part \( b \) of the complex number \( (a +ib) \).
      * @return The argument of the complex number.
      * @see Math#atan2(double, double)
      */
@@ -156,8 +158,8 @@ public final class ComplexFunctions {
      *
      * <p>Note: A complex number with at least one infinite part is regarded
      * as an infinity (even if its other part is a NaN).
-     * @param real Real part \( a \) of the complex number \(a +ib \).
-     * @param imaginary Imaginary part \( b \) of the complex number \(a +ib \).
+     * @param real Real part \( a \) of the complex number \( (a +ib) \).
+     * @param imaginary Imaginary part \( b \) of the complex number \( (a +ib) \).
      * @return {@code true} if the complex number contains an infinite value.
      * @see Double#isInfinite(double)
      */
@@ -204,16 +206,16 @@ public final class ComplexFunctions {
      * ACM Transactions on Mathematical Software, Vol 20, No 2, pp 215-244.
      * </blockquote>
      *
-     * @param real Real part \( a \) of the complex number \(a +ib \).
-     * @param imaginary Imaginary part \( b \) of the complex number \(a +ib \).
-     * @param constructor Terminating consumer for the complex result, used to construct a result of type {@code R}.
-     * @param <R> the object type produced by the supplied constructor.
-     * @return The natural logarithm of the complex number.
+     * @param real Real part \( a \) of the complex number \( (a +ib \).
+     * @param imaginary Imaginary part \( b \) of the complex number \( (a +ib \).
+     * @param action Consumer for the natural logarithm of the complex number.
+     * @param <R> the object taken by the supplied action.
+     * @return the object returned by the supplied action.
      * @see Math#log(double)
      * @see <a href="http://functions.wolfram.com/ElementaryFunctions/Log/">Log</a>
      */
-    public static <R> R log(double real, double imaginary, ComplexResult<R> constructor) {
-        return log(Math::log, HALF, LN_2, real, imaginary, constructor);
+    public static <R> R log(double real, double imaginary, ComplexSink<R> action) {
+        return log(Math::log, HALF, LN_2, real, imaginary, action);
     }
 
     /**
@@ -232,15 +234,15 @@ public final class ComplexFunctions {
      *
      * <p>where \( |z| \) is the absolute and \( \arg(z) \) is the argument.
      *
-     * @param real Real part \( a \) of the complex number \(a +ib \).
-     * @param imaginary Imaginary part \( b \) of the complex number \(a +ib \).
-     * @param constructor Terminating consumer for the complex result, used to construct a result of type {@code R}.
-     * @param <R> the object type produced by the supplied constructor.
-     * @return The base 10 logarithm of the complex number.
+     * @param real Real part \( a \) of the complex number \( (a +ib \).
+     * @param imaginary Imaginary part \( b \) of the complex number \( (a +ib \).
+     * @param action Consumer for the natural logarithm of the complex number.
+     * @param <R> the object taken by the supplied action.
+     * @return the object returned by the supplied action.
      * @see Math#log10(double)
      */
-    public static <R> R log10(double real, double imaginary, ComplexResult<R> constructor) {
-        return log(Math::log10, LOG_10E_O_2, LOG10_2, real, imaginary, constructor);
+    public static <R> R log10(double real, double imaginary, ComplexSink<R> action) {
+        return log(Math::log10, LOG_10E_O_2, LOG10_2, real, imaginary, action);
     }
 
     /**
@@ -258,26 +260,26 @@ public final class ComplexFunctions {
      * @param log Log function.
      * @param logOfeOver2 The log function applied to e, then divided by 2.
      * @param logOf2 The log function applied to 2.
-     * @param real Real part \( a \) of the complex number \(a +ib \).
-     * @param imaginary Imaginary part \( b \) of the complex number \(a +ib \).
-     * @param constructor Terminating consumer for the complex result, used to construct a result of type {@code R}.
-     * @param <R> the object type produced by the supplied constructor.
-     * @return The logarithm of the complex number.
+     * @param real Real part \( a \) of the complex number \( (a +ib \).
+     * @param imaginary Imaginary part \( b \) of the complex number \( (a +ib \).
+     * @param action Consumer for the natural logarithm of the complex number.
+     * @param <R> the object taken by the supplied action.
+     * @return the object returned by the supplied action.
      */
     private static <R> R log(DoubleUnaryOperator log,
         double logOfeOver2,
         double logOf2,
         double real,
         double imaginary,
-        ComplexResult<R> constructor) {
+        ComplexSink<R> action) {
 
         // Handle NaN
         if (Double.isNaN(real) || Double.isNaN(imaginary)) {
             // Return NaN unless infinite
             if (isInfinite(real, imaginary)) {
-                return constructor.apply(Double.POSITIVE_INFINITY, Double.NaN);
+                return action.apply(Double.POSITIVE_INFINITY, Double.NaN);
             }
-            return constructor.apply(Double.NaN, Double.NaN);
+            return action.apply(Double.NaN, Double.NaN);
         }
 
         // Returns the real part:
@@ -297,8 +299,8 @@ public final class ComplexFunctions {
 
         if (x == 0) {
             // Handle zero: raises the ‘‘divide-by-zero’’ floating-point exception.
-            return constructor.apply(Double.NEGATIVE_INFINITY,
-                Complex.negative(real) ? Math.copySign(Math.PI, imaginary) : imaginary);
+            return action.apply(Double.NEGATIVE_INFINITY,
+                negative(real) ? Math.copySign(Math.PI, imaginary) : imaginary);
         }
 
         double re;
@@ -312,7 +314,7 @@ public final class ComplexFunctions {
 
         if (x > HALF && x < ROOT2) {
             // x^2+y^2 close to 1. Use log1p(x^2+y^2 - 1) / 2.
-            re = Math.log1p(Complex.x2y2m1(x, y)) * logOfeOver2;
+            re = Math.log1p(x2y2m1(x, y)) * logOfeOver2;
         } else {
             // Check for over/underflow in |z|
             // When scaling:
@@ -321,9 +323,9 @@ public final class ComplexFunctions {
             re = 0;
             if (x > Double.MAX_VALUE / 2) {
                 // Potential overflow.
-                if (Complex.isPosInfinite(x)) {
+                if (isPosInfinite(x)) {
                     // Handle infinity
-                    return constructor.apply(x, arg(real, imaginary));
+                    return action.apply(x, arg(real, imaginary));
                 }
                 // Scale down.
                 x /= 2;
@@ -334,7 +336,7 @@ public final class ComplexFunctions {
                 // Potential underflow.
                 if (y == 0) {
                     // Handle real only number
-                    return constructor.apply(log.applyAsDouble(x), arg(real, imaginary));
+                    return action.apply(log.applyAsDouble(x), arg(real, imaginary));
                 }
                 // Scale up sub-normal numbers to make them normal by scaling by 2^54,
                 // i.e. more than the mantissa digits.
@@ -347,7 +349,7 @@ public final class ComplexFunctions {
         }
 
         // All ISO C99 edge cases for the imaginary are satisfied by the Math library.
-        return constructor.apply(re, arg(real, imaginary));
+        return action.apply(re, arg(real, imaginary));
     }
 
     /**
@@ -441,7 +443,7 @@ public final class ComplexFunctions {
         // insensitive to the sign bit. Thus use of Math.abs(double) is only in edge-case
         // branches.
         //
-        // 5. The exponent different to ignore the smaller component has changed from 60 to 54.
+        // 5. The exponent difference to ignore the smaller component has changed from 60 to 54.
         //
         // Original comments from fdlibm are in c style: /* */
         // Extra comments added for reference.
@@ -561,13 +563,13 @@ public final class ComplexFunctions {
         // xxLow = Math.fma(x, x, -xx)
         // yyLow = Math.fma(y, y, -yy)
         // Dekker mul12
-        final double xHigh = Complex.splitHigh(x);
+        final double xHigh = splitHigh(x);
         final double xLow = x - xHigh;
-        final double xxLow = Complex.squareLow(xLow, xHigh, xx);
+        final double xxLow = squareLow(xLow, xHigh, xx);
         // Dekker mul12
-        final double yHigh = Complex.splitHigh(y);
+        final double yHigh = splitHigh(y);
         final double yLow = y - yHigh;
-        final double yyLow = Complex.squareLow(yLow, yHigh, yy);
+        final double yyLow = squareLow(yLow, yHigh, yy);
         // Dekker add2
         final double r = xx + yy;
         // Note: The order is important. Assume xx > yy and drop Dekker's conditional
@@ -580,5 +582,264 @@ public final class ComplexFunctions {
         // That adds 7 multiply, 1 division and 19 additions doubling the cost
         // and reducing error to < 0.5 ulp for the final sqrt.
         return xx - r + yy + yyLow + xxLow + r;
+    }
+
+    /**
+     * Compute {@code x^2 + y^2 - 1} in high precision.
+     * Assumes that the values x and y can be multiplied without overflow; that
+     * {@code x >= y}; and both values are positive.
+     *
+     * @param x the x value
+     * @param y the y value
+     * @return {@code x^2 + y^2 - 1}.
+     */
+    //TODO - make it private in future
+    static double x2y2m1(double x, double y) {
+        // Hull et al used (x-1)*(x+1)+y*y.
+        // From the paper on page 236:
+
+        // If x == 1 there is no cancellation.
+
+        // If x > 1, there is also no cancellation, but the argument is now accurate
+        // only to within a factor of 1 + 3 EPSILSON (note that x – 1 is exact),
+        // so that error = 3 EPSILON.
+
+        // If x < 1, there can be serious cancellation:
+
+        // If 4 y^2 < |x^2 – 1| the cancellation is not serious ... the argument is accurate
+        // only to within a factor of 1 + 4 EPSILSON so that error = 4 EPSILON.
+
+        // Otherwise there can be serious cancellation and the relative error in the real part
+        // could be enormous.
+
+        final double xx = x * x;
+        final double yy = y * y;
+        // Modify to use high precision before the threshold set by Hull et al.
+        // This is to preserve the monotonic output of the computation at the switch.
+        // Set the threshold when x^2 + y^2 is above 0.5 thus subtracting 1 results in a number
+        // that can be expressed with a higher precision than any number in the range 0.5-1.0
+        // due to the variable exponent used below 0.5.
+        if (x < 1 && xx + yy > 0.5) {
+            // Large relative error.
+            // This does not use o.a.c.numbers.LinearCombination.value(x, x, y, y, 1, -1).
+            // It is optimised knowing that:
+            // - the products are squares
+            // - the final term is -1 (which does not require split multiplication and addition)
+            // - The answer will not be NaN as the terms are not NaN components
+            // - The order is known to be 1 > |x| >= |y|
+            // The squares are computed using a split multiply algorithm and
+            // the summation using an extended precision summation algorithm.
+
+            // Split x and y as one 26 bits number and one 27 bits number
+            final double xHigh = splitHigh(x);
+            final double xLow  = x - xHigh;
+            final double yHigh = splitHigh(y);
+            final double yLow  = y - yHigh;
+
+            // Accurate split multiplication x * x and y * y
+            final double x2Low = squareLow(xLow, xHigh, xx);
+            final double y2Low = squareLow(yLow, yHigh, yy);
+
+            return sumx2y2m1(xx, x2Low, yy, y2Low);
+        }
+        return (x - 1) * (x + 1) + yy;
+    }
+
+    /**
+     * Implement Dekker's method to split a value into two parts. Multiplying by (2^s + 1) create
+     * a big value from which to derive the two split parts.
+     * <pre>
+     * c = (2^s + 1) * a
+     * a_big = c - a
+     * a_hi = c - a_big
+     * a_lo = a - a_hi
+     * a = a_hi + a_lo
+     * </pre>
+     *
+     * <p>The multiplicand must be odd allowing a p-bit value to be split into
+     * (p-s)-bit value {@code a_hi} and a non-overlapping (s-1)-bit value {@code a_lo}.
+     * Combined they have (p􏰔-1) bits of significand but the sign bit of {@code a_lo}
+     * contains a bit of information.
+     *
+     * @param a Value.
+     * @return the high part of the value.
+     * @see <a href="https://doi.org/10.1007/BF01397083">
+     * Dekker (1971) A floating-point technique for extending the available precision</a>
+     */
+    static double splitHigh(double a) {
+        final double c = MULTIPLIER * a;
+        return c - (c - a);
+    }
+
+    /**
+     * Compute the round-off from the square of a split number with {@code low} and {@code high}
+     * components. Uses Dekker's algorithm for split multiplication modified for a square product.
+     *
+     * <p>Note: This is candidate to be replaced with {@code Math.fma(x, x, -x * x)} to compute
+     * the round-off from the square product {@code x * x}. This would remove the requirement
+     * to compute the split number and make this method redundant. {@code Math.fma} requires
+     * JDK 9 and FMA hardware support.
+     *
+     * @param low Low part of number.
+     * @param high High part of number.
+     * @param square Square of the number.
+     * @return <code>low * low - (((product - high * high) - low * high) - high * low)</code>
+     * @see <a href="http://www-2.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps">
+     * Shewchuk (1997) Theorum 18</a>
+     */
+    static double squareLow(double low, double high, double square) {
+        final double lh = low * high;
+        return low * low - (((square - high * high) - lh) - lh);
+    }
+
+    /**
+     * Compute the round-off from the sum of two numbers {@code a} and {@code b} using
+     * Dekker's two-sum algorithm. The values are required to be ordered by magnitude:
+     * {@code |a| >= |b|}.
+     *
+     * @param a First part of sum.
+     * @param b Second part of sum.
+     * @param x Sum.
+     * @return <code>b - (x - a)</code>
+     * @see <a href="http://www-2.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps">
+     * Shewchuk (1997) Theorum 6</a>
+     */
+    private static double fastSumLow(double a, double b, double x) {
+        // x = a + b
+        // bVirtual = x - a
+        // y = b - bVirtual
+        return b - (x - a);
+    }
+
+    /**
+     * Compute the round-off from the sum of two numbers {@code a} and {@code b} using
+     * Knuth's two-sum algorithm. The values are not required to be ordered by magnitude.
+     *
+     * @param a First part of sum.
+     * @param b Second part of sum.
+     * @param x Sum.
+     * @return <code>(a - (x - (x - a))) + (b - (x - a))</code>
+     * @see <a href="http://www-2.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps">
+     * Shewchuk (1997) Theorum 7</a>
+     */
+    private static double sumLow(double a, double b, double x) {
+        // x = a + b
+        // bVirtual = x - a
+        // aVirtual = x - bVirtual
+        // bRoundoff = b - bVirtual
+        // aRoundoff = a - aVirtual
+        // y = aRoundoff + bRoundoff
+        final double bVirtual = x - a;
+        return (a - (x - bVirtual)) + (b - bVirtual);
+    }
+
+    /**
+     * Sum x^2 + y^2 - 1. It is assumed that {@code y <= x < 1}.
+     *
+     * <p>Implement Shewchuk's expansion-sum algorithm: [x2Low, x2High] + [-1] + [y2Low, y2High].
+     *
+     * @param x2High High part of x^2.
+     * @param x2Low Low part of x^2.
+     * @param y2High High part of y^2.
+     * @param y2Low Low part of y^2.
+     * @return x^2 + y^2 - 1
+     * @see <a href="http://www-2.cs.cmu.edu/afs/cs/project/quake/public/papers/robust-arithmetic.ps">
+     * Shewchuk (1997) Theorum 12</a>
+     */
+    private static double sumx2y2m1(double x2High, double x2Low, double y2High, double y2Low) {
+        // Let e and f be non-overlapping expansions of components of length m and n.
+        // The following algorithm will produce a non-overlapping expansion h where the
+        // sum h_i = e + f and components of h are in increasing order of magnitude.
+        // Expansion-sum proceeds by a grow-expansion of the first part from one expansion
+        // into the other, extending its length by 1. The process repeats for the next part
+        // but the grow-expansion starts at the previous merge position + 1.
+        // Thus expansion-sum requires mn two-sum operations to merge length m into length n
+        // resulting in length m+n-1.
+
+        // Variables numbered from 1 as per Figure 7 (p.12). The output expansion h is placed
+        // into e increasing its length for each grow expansion.
+
+        // We have two expansions for x^2 and y^2 and the whole number -1.
+        // Expecting (x^2 + y^2) close to 1 we generate first the intermediate expansion
+        // (x^2 - 1) moving the result away from 1 where there are sparse floating point
+        // representations. This is then added to a similar magnitude y^2. Leaving the -1
+        // until last suffers from 1 ulp rounding errors more often and the requirement
+        // for a distillation sum to reduce rounding error frequency.
+
+        // Note: Do not use the alternative fast-expansion-sum of the parts sorted by magnitude.
+        // The parts can be ordered with a single comparison into:
+        // [y2Low, (y2High|x2Low), x2High, -1]
+        // The fast-two-sum saves 1 fast-two-sum and 3 two-sum operations (21 additions) and
+        // adds a penalty of a single branch condition.
+        // However the order in not "strongly non-overlapping" and the fast-expansion-sum
+        // output will not be strongly non-overlapping. The sum of the output has 1 ulp error
+        // on random cis numbers approximately 1 in 160 events. This can be removed by a
+        // distillation two-sum pass over the final expansion as a cost of 1 fast-two-sum and
+        // 3 two-sum operations! So we use the expansion sum with the same operations and
+        // no branches.
+
+        double q = x2Low - 1;
+        double e1 = fastSumLow(-1, x2Low, q);
+        double e3 = q + x2High;
+        double e2 = sumLow(q, x2High, e3);
+
+        final double f1 = y2Low;
+        final double f2 = y2High;
+
+        // Grow expansion of f1 into e
+        q = f1 + e1;
+        e1 = sumLow(f1, e1, q);
+        double p = q + e2;
+        e2 = sumLow(q, e2, p);
+        double e4 = p + e3;
+        e3 = sumLow(p, e3, e4);
+
+        // Grow expansion of f2 into e (only required to start at e2)
+        q = f2 + e2;
+        e2 = sumLow(f2, e2, q);
+        p = q + e3;
+        e3 = sumLow(q, e3, p);
+        final double e5 = p + e4;
+        e4 = sumLow(p, e4, e5);
+
+        // Final summation:
+        // The sum of the parts is within 1 ulp of the true expansion value e:
+        // |e - sum| < ulp(sum).
+        // To achieve the exact result requires iteration of a distillation two-sum through
+        // the expansion until convergence, i.e. no smaller term changes higher terms.
+        // This requires (n-1) iterations for length n. Here we neglect this as
+        // although the method is not ensured to be exact is it robust on random
+        // cis numbers.
+        return e1 + e2 + e3 + e4 + e5;
+    }
+
+    /**
+     * Check that a value is negative. It must meet all the following conditions:
+     * <ul>
+     *  <li>it is not {@code NaN},</li>
+     *  <li>it is negative signed,</li>
+     * </ul>
+     *
+     * <p>Note: This is true for negative zero.</p>
+     *
+     * @param d Value.
+     * @return {@code true} if {@code d} is negative.
+     */
+    //TODO - make private in future
+    static boolean negative(double d) {
+        return d < 0 || Double.doubleToLongBits(d) == NEGATIVE_ZERO_LONG_BITS;
+    }
+
+    /**
+     * Check that a value is positive infinity. Used to replace {@link Double#isInfinite()}
+     * when the input value is known to be positive (i.e. in the case where it has been
+     * set using {@link Math#abs(double)}).
+     *
+     * @param d Value.
+     * @return {@code true} if {@code d} is +inf.
+     */
+    //TODO - make private in future
+    static boolean isPosInfinite(double d) {
+        return d == Double.POSITIVE_INFINITY;
     }
 }
