+++ commons-numbers-complex/src/main/java/org/apache/commons/numbers/complex/ComplexSink.java
@@ -18,22 +18,22 @@
 package org.apache.commons.numbers.complex;
 
 /**
- * Represents a Terminating consumer for the complex result, used to construct a result of type {@code R}.
- * The operation may return an object representation of the complex result.
+ * Represents a data sink for a complex number \( (a + i b) \)
+ * Operations return a result of type {@code R}.
  *
  * <p>This is a functional interface whose functional method is
  * {@link #apply(double, double)}.
  *
- * @param <R> The type of the complex result
+ * @param <R> The type of the result
  * @since 1.1
  */
 @FunctionalInterface
-public interface ComplexResult<R> {
+public interface ComplexSink<R> {
 
     /**
      * Represents a function that accepts real and imaginary part of complex number and returns an object.
-     * @param real Real part \( a \) of the complex number \(a +ib \).
-     * @param imaginary Imaginary part \( b \) of the complex number \(a +ib \).
+     * @param real Real part \( a \) of the complex number \( (a +ib) \).
+     * @param imaginary Imaginary part \( b \) of the complex number \( (a +ib) \).
      * @return R the object encapsulating the complex result
      */
     R apply(double real, double imaginary);
